freeSeats updated everywhere

+++ budimirovic.ana.eventsApp/EventsApp/app/src/main/java/com/example/eventsapp/Event.java
@@ -147,7 +147,7 @@ public class Event {
     public double getAverageRating() {
         return averageRating;
     }
-    public double getNumberOfAttendees() {
+    public int getNumberOfAttendees() {
         return numbOfAttendees;
     }
 }
